requirements.txt
alembic==1.13.0
aniso8601==9.0.1
apispec==6.3.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
DateTime==5.4
Flask==3.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
flask-smorest==0.42.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.2
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.2
passlib==1.7.4
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
setuptools==69.0.2
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
webargs==8.3.0
Werkzeug==3.0.1
zipp==3.17.0
zope.interface==6.1